Fix seed setting error

+++ main.py
@@ -58,8 +58,8 @@ while GaPipeline.gen_count <= num_gen:
         if surrogate_enabled:
             all_subsurrogate_metrics = GaPipeline.prepare_surrogate()
     
-    random.seed(SEED*(GaPipeline.gen_count+1))
-    np.random.seed(SEED*(GaPipeline.gen_count+1))
+    random.seed(int(SEED*(GaPipeline.gen_count+1)))
+    np.random.seed(int(SEED*(GaPipeline.gen_count+1)))
     
     if not GaPipeline.attempt_resume:
         elites = GaPipeline.update_elite_pool() # elites are selected from existing elite pool and current pop
